Removing eslint/compat + explicit import of globals ## Summary
Optional summary of the PR here. The GitHub release description is created from this comment so keep it nice and descriptive.
Remember to remove sections that you don't need or use.

If it does not make sense to have a summary, you can take that out as well.
You would typically keep the summary only if there is a general theme change that can be summarized. If you find
yourself saying the same thing as any of the bullet points in a slightly different way; then this section is not
needed.

### Added

- Describe the added features

### Changed

- Describe the outwards facing code change

### Fixed

- Describe the fix and the bug

### Removed

- Describe what was removed and why

### Security

- Describe the security issue and the fix

### Deprecated

- Describe the part of the code being deprecated and why